+++ demo-app/app/src/main/java/com/launchkey/android/whitelabel/demo/app/DemoApplication.java
@@ -2,21 +2,29 @@ package com.launchkey.android.whitelabel.demo.app;
 
 import android.app.Notification;
 import android.app.PendingIntent;
-import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.app.NotificationManagerCompat;
-import android.support.v4.content.LocalBroadcastManager;
+import android.util.Log;
 
+import com.launchkey.android.authenticator.sdk.AuthenticatorApplication;
+import com.launchkey.android.authenticator.sdk.AuthenticatorConfig;
+import com.launchkey.android.authenticator.sdk.AuthenticatorManager;
+import com.launchkey.android.authenticator.sdk.DeviceLinkedEventCallback;
+import com.launchkey.android.authenticator.sdk.DeviceUnlinkedEventCallback;
+import com.launchkey.android.authenticator.sdk.auth.AuthRequest;
+import com.launchkey.android.authenticator.sdk.auth.AuthRequestManager;
+import com.launchkey.android.authenticator.sdk.auth.event.GetAuthRequestEventCallback;
+import com.launchkey.android.authenticator.sdk.device.Device;
+import com.launchkey.android.authenticator.sdk.error.BaseError;
 import com.launchkey.android.whitelabel.demo.R;
 import com.launchkey.android.whitelabel.demo.ui.activity.ListDemoActivity;
-import com.launchkey.android.whitelabel.sdk.WhiteLabelApplication;
-import com.launchkey.android.whitelabel.sdk.WhiteLabelConfig;
-import com.launchkey.android.whitelabel.sdk.WhiteLabelManager;
 
-public class DemoApplication extends WhiteLabelApplication {
+import java.util.Locale;
+
+public class DemoApplication extends AuthenticatorApplication {
+
+    public static final String TAG = DemoApplication.class.getSimpleName();
 
     private static final int NOTIFICATION_ID = 100;
 
@@ -24,38 +32,54 @@ public class DemoApplication extends WhiteLabelApplication {
 
     @Override
     public void onCreate() {
-        init();
-        prepareForRequestNotifications();
+        initialize();
         super.onCreate();
     }
 
-    private void init() {
-        int keyPairSizeBits = WhiteLabelConfig.Builder.KEYSIZE_MEDIUM;
+    private void initialize() {
+        int keyPairSizeBits = AuthenticatorConfig.Builder.KEYSIZE_MINIMUM;
+        //int keyPairSizeBits = WhiteLabelConfig.Builder.KEYSIZE_MEDIUM;
         //keyPairSizeBits = 3072; //Could also assign the actual value in bits.
 
-        WhiteLabelConfig c = new WhiteLabelConfig.Builder(this, R.string.whitelabel_sdk_key)
-                .endpoint(getString(R.string.whitelabel_sdk_endpoint))
-                .keyPairSize(keyPairSizeBits)
-                .themeColorPrimaryRes(R.color.demo_primary)
-                .themeColorPrimaryTextAndIconsRes(R.color.demo_primaryTextIcons)
-                .themeColorSecondaryRes(R.color.demo_accent)
-                .themeColorBackgroundsRes(R.color.demo_backgrounds)
-                .build();
+        final AuthenticatorManager manager = AuthenticatorManager.getInstance();
 
-        WhiteLabelManager.getInstance().init(c);
-    }
+        manager.initialize(
+                        new AuthenticatorConfig.Builder(this, R.string.authenticator_sdk_key)
+                                .keyPairSize(keyPairSizeBits)
+                                .theme(R.style.DemoAppTheme)
+                                .build());
 
-    private void prepareForRequestNotifications() {
+        manager.registerForEvents(
 
-        BroadcastReceiver requestReceiver = new BroadcastReceiver() {
-            @Override
-            public void onReceive(Context context, Intent intent) {
-                notifyOfRequest();
-            }
-        };
+                new DeviceLinkedEventCallback() {
+                    @Override
+                    public void onEventResult(boolean b, BaseError baseError, Device device) {
+                        Log.i(TAG, String.format(Locale.getDefault(), "Link-event=%b Device-name=%s", b, device.getName()));
+                    }
+                },
 
-        IntentFilter requestFilter = new IntentFilter(WhiteLabelManager.ACTION_EVENT_REQUEST_INCOMING);
-        LocalBroadcastManager.getInstance(this).registerReceiver(requestReceiver, requestFilter);
+                new DeviceUnlinkedEventCallback() {
+                    @Override
+                    public void onEventResult(boolean b, BaseError baseError, Object o) {
+                        Log.i(TAG, String.format(Locale.getDefault(), "Unlink-event=%b error=%s", b, baseError));
+                    }
+                }
+        );
+
+        if (manager.isDeviceLinked()) {
+
+            final AuthRequestManager arm = AuthRequestManager.getInstance(this);
+            arm.registerForEvents(new GetAuthRequestEventCallback() {
+
+                @Override
+                public void onEventResult(boolean successful, BaseError error, AuthRequest authRequest) {
+
+                    if (authRequest != null) {
+                        notifyOfRequest();
+                    }
+                }
+            });
+        }
     }
 
     private void notifyOfRequest() {