+++ demo-app/app/src/main/java/com/launchkey/android/whitelabel/demo/ui/activity/BaseDemoActivity.java
@@ -4,24 +4,24 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
+import com.launchkey.android.authenticator.sdk.AuthenticatorManager;
 import com.launchkey.android.whitelabel.demo.ui.BaseDemoView;
-import com.launchkey.android.whitelabel.sdk.WhiteLabelManager;
 
 /**
  * Created by armando on 7/8/16.
  */
 public class BaseDemoActivity extends AppCompatActivity implements BaseDemoView {
 
-    private WhiteLabelManager mWhiteLabelManager;
+    private AuthenticatorManager mAuthenticatorManager;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        mWhiteLabelManager = WhiteLabelManager.getInstance();
+        mAuthenticatorManager = AuthenticatorManager.getInstance();
     }
 
     @Override
-    public WhiteLabelManager getWhiteLabelManager() {
-        return mWhiteLabelManager;
+    public AuthenticatorManager getAuthenticatorManager() {
+        return mAuthenticatorManager;
     }
 }